In ``summit/support``, exchange ip between ospl-azar1 and ospl-love2, increase master priority of ospl-azar1 to 202.
This is to fix an an issue encoutered during the summit deployment were the dds network disapeared momentarily from azar1 and I deployed the main daemon of love2 with the same configuration from azar1.
This modificatin will ensure azar1 daemon will become main and that its IP won't colide with the one now running in love2.
After I deploy this on azar1 I will redeploy the daemon on love2, thus removing it's aligner configuration.